Revert "refactor:! merged sync admin and non-admin clients"
This reverts commit 8717a88e068e0b94ba150e687d7c30a7943763a4.
