+++ docs/admin.rst
@@ -1,0 +1,10 @@
+AdminClient
+=====================
+
+This class is the admin client, designed to conduct AVS administrative operation such as creating indexes, querying index information, and dropping indexes.
+
+
+.. autoclass:: aerospike_vector_search.admin.Client
+   :members:
+   :undoc-members:
+   :show-inheritance:
