+++ src/aerospike_vector_search/__init__.py
@@ -1,4 +1,5 @@
 from .client import Client
+from .admin import Client as AdminClient
 from .types import (
     HostPort,
     Key,
