+++ tests/rbac/sync/test_admin_client_get_user.py
@@ -22,12 +22,12 @@ class get_user_test_case:
         ),
     ],
 )
-def test_get_user(session_rbac_client, test_case):
-    session_rbac_client.add_user(
+def test_get_user(session_rbac_admin_client, test_case):
+    session_rbac_admin_client.add_user(
         username=test_case.username, password=test_case.password, roles=None
     )
 
-    result = session_rbac_client.get_user(username=test_case.username)
+    result = session_rbac_admin_client.get_user(username=test_case.username)
 
     assert result.username == test_case.username
