+++ tests/standard/test_vector_client_search_by_key.py
@@ -265,10 +265,11 @@ class vector_search_by_key_test_case:
 )
 def test_vector_search_by_key(
     session_vector_client,
+    session_admin_client,
     test_case,
 ):
     
-    session_vector_client.index_create(
+    session_admin_client.index_create(
         namespace=test_case.search_namespace,
         name=test_case.index_name,
         vector_field=test_case.vector_field,
@@ -297,7 +298,7 @@ def test_vector_search_by_key(
         )
     
     wait_for_index(
-        admin_client=session_vector_client,
+        admin_client=session_admin_client,
         namespace=test_case.search_namespace,
         index=test_case.index_name,
     )
@@ -323,7 +324,7 @@ def test_vector_search_by_key(
             key=key,
         )
 
-    session_vector_client.index_drop(
+    session_admin_client.index_drop(
         namespace=test_case.search_namespace,
         name=test_case.index_name,
     )
@@ -331,9 +332,10 @@ def test_vector_search_by_key(
 
 def test_vector_search_by_key_different_namespaces(
     session_vector_client,
+    session_admin_client,
 ):
     
-    session_vector_client.index_create(
+    session_admin_client.index_create(
         namespace="index_storage",
         name="diff_ns_idx",
         vector_field="vec",
@@ -372,7 +374,7 @@ def test_vector_search_by_key_different_namespaces(
     )
     
     wait_for_index(
-        admin_client=session_vector_client,
+        admin_client=session_admin_client,
         namespace="index_storage",
         index="diff_ns_idx",
     )
@@ -413,7 +415,7 @@ def test_vector_search_by_key_different_namespaces(
         key="search_for",
     )
 
-    session_vector_client.index_drop(
+    session_admin_client.index_drop(
         namespace="index_storage",
         name="diff_ns_idx",
     )
